+++ app/lib/auth.prefetches.ts
