+++ app/routes/home/components/posts.card.tsx
@@ -17,7 +17,7 @@ import { ClientOnly } from "remix-utils/client-only";
 import { authClient } from "~/lib/auth";
 import { cn, formatTime } from "~/lib/utils";
 
-import axiosInstance from "~/lib/axios-instance";
+import axiosInstance from "~/lib/axios.instance";
 import queryClient from "~/lib/query/query-client";
 
 import type { Post, User } from "~/lib/types/shared";
