+++ app/routes/home/index.tsx
@@ -6,7 +6,7 @@ import { useWindowVirtualizer } from "@tanstack/react-virtual";
 import React from "react";
 import { useSearchParams } from "react-router";
 
-import axiosInstance from "~/lib/axios-instance";
+import axiosInstance from "~/lib/axios.instance";
 
 import type { Post } from "~/lib/types/shared";
