+++ app/routes/post/components/posts.card.tsx
@@ -16,7 +16,7 @@ import {
 
 import { cn, formatTime } from "~/lib/utils";
 
-import axiosInstance from "~/lib/axios-instance";
+import axiosInstance from "~/lib/axios.instance";
 import queryClient from "~/lib/query/query-client";
 
 import type { Session } from "~/lib/auth";
